:bookmark: Version 11.8.0
:arrow_up: Upgraded MFXCore to version 11.1.4

:boom: Implemented virtualized table as VirtualTable as well as its paginated version as PaginatedVirtualTable

Controls Package
:sparkles: Imported BoundLabel from MaterialFX as it was needed to implement DefaultTableColumnSkin

Flow Package
:recycle: Minor refactors to FlowState and ViewportManager
:recycle: VirtualFlowSkin: process items change before computing the new estimated size

Grid Package
:memo: Minor fixes to documentation here and there

Tests
:sparkles: Added new bean to test the table, using Faker to generate random data
:sparkles: Added convenient method to Constraint class to generate a constraint on a list
:sparkles: Added new dialog to get a generic choice from the user
:sparkles: Added new method to Utils to get all the items at the given indexes from a given list
:fire: Remove tests on viewport sizes for paginated variants

Resources
:recycle: Common.css: make combo popups taller
:bug: GridTest.fxml: remove minHeight constraint from HBox as it was causing weird issues with layout

+++ src/test/java/interactive/cells/CellsDebugger.java
@@ -47,7 +47,7 @@ public class CellsDebugger {
 		));
 	}
 
-	public static Color getColor(Integer item, float opacity) {
+	public static Color getColor(Integer item, double opacity) {
 		return dColors.computeIfAbsent(item, i -> Color.color(
 				RandomUtils.random.nextFloat(),
 				RandomUtils.random.nextFloat(),
@@ -60,11 +60,11 @@ public class CellsDebugger {
 		dColors.clear();
 	}
 
-	public static void randBackground(Region r, float opacity, Integer item) {
+	public static void randBackground(Region r, double opacity, Integer item) {
 		StyleUtils.setBackground(r, getColor(item, opacity));
 	}
 
-	public static void randBackground(Region r, float opacity, Integer item, double radius, double inset) {
+	public static void randBackground(Region r, double opacity, Integer item, double radius, double inset) {
 		StyleUtils.setBackground(r, getColor(item, opacity), CornerRadiusFactory.all(radius), InsetsFactory.all(inset));
 	}
